+++ packages/meteor/src/index.ts
@@ -2,6 +2,7 @@ export * from './MeteorCall';
 export * from './MeteorDataFetcher';
 export * from './MeteorSubscription';
 export * from './MeteorDependencies';
+export * from './MeteorPublication';
 export * from './MeteorSubscriptionImpl';
 export * from './MeteorTableOperations';
 export * from './MeteorTableFetcher';